extract match summary stuff
